bring parameter cert over
